+++ HotSource/src/main/java/hotsource/domain/Ordered.java
@@ -7,7 +7,7 @@ import lombok.Data;
 
 @Data
 public class Ordered {
-	private int order_id;
+	private long order_id;
 	private String order_no;
 	private Timestamp order_date;
 	private String order_state;
